Revert "Merge pull request #682 from bugsnag/integration/javascript-span-access"
This reverts commit aa228b45ff00625fd624ac77ec848d56414e59fb, reversing
changes made to daeb74d97a118f9f851a5da229d31f94511d3660.
