+++ packages/platforms/browser/tests/clock.test.ts
@@ -164,53 +164,4 @@ describe('Browser Clock', () => {
       expect(clock.convert(new Date())).toEqual(0)
     })
   })
-
-  describe('clock.toUnixNanoseconds()', () => {
-    it('converts performance time to unix nanoseconds', () => {
-      const timeOrigin = new Date('2023-01-02T00:00:00.000Z')
-      jest.setSystemTime(timeOrigin)
-
-      const clock = createClock(new PerformanceFake(), new ControllableBackgroundingListener())
-
-      jest.advanceTimersByTime(250)
-
-      const performanceTime = clock.now()
-      const unixNanoseconds = clock.toUnixNanoseconds(performanceTime)
-
-      // Expected: timeOrigin + 250ms in nanoseconds
-      const expectedNanoseconds = (timeOrigin.getTime() + 250) * 1000000
-      expect(unixNanoseconds).toBe(expectedNanoseconds)
-    })
-  })
-
-  describe('clock.fromUnixNanoseconds()', () => {
-    it('converts unix nanoseconds back to performance time', () => {
-      const timeOrigin = new Date('2023-01-02T00:00:00.000Z')
-      jest.setSystemTime(timeOrigin)
-
-      const clock = createClock(new PerformanceFake(), new ControllableBackgroundingListener())
-
-      // Test with a specific unix nanoseconds value 123ms after timeOrigin
-      const unixNanoseconds = 1672617600123000000 // 2023-01-02T00:00:00.123Z in nanoseconds
-      const performanceTime = clock.fromUnixNanoseconds(unixNanoseconds)
-
-      expect(typeof performanceTime).toBe('number')
-      expect(performanceTime).toEqual(123) // 123ms after timeOrigin
-    })
-
-    it('is the inverse of toUnixNanoseconds', () => {
-      const timeOrigin = new Date('2023-01-02T00:00:00.000Z')
-      jest.setSystemTime(timeOrigin)
-
-      const clock = createClock(new PerformanceFake(), new ControllableBackgroundingListener())
-
-      jest.advanceTimersByTime(123)
-
-      const originalTime = clock.now()
-      const unixNanoseconds = clock.toUnixNanoseconds(originalTime)
-      const roundTripTime = clock.fromUnixNanoseconds(unixNanoseconds)
-
-      expect(roundTripTime).toBe(originalTime)
-    })
-  })
 })
