--- packages/plugin-react-native-span-access/android/src/main/java/com/bugsnag/reactnative/performance/nativespans/JavascriptSpanTransaction.java
@@ -1,13 +1,0 @@
-package com.bugsnag.reactnative.performance.nativespans;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-
-public interface JavascriptSpanTransaction {
-    JavascriptSpanTransaction end();
-    JavascriptSpanTransaction end(long endTime);
-
-    JavascriptSpanTransaction setAttribute(@NonNull String key, @Nullable Object value);
-
-    void commit(@Nullable OnRemoteSpanUpdatedCallback callback);
-}
