--- packages/plugin-react-native-span-access/android/src/main/java/com/bugsnag/reactnative/performance/nativespans/OnRemoteSpanUpdatedCallback.java
@@ -1,5 +1,0 @@
-package com.bugsnag.reactnative.performance.nativespans;
-
-public interface OnRemoteSpanUpdatedCallback {
-  void onRemoteSpanUpdated(boolean updateResult);
-}
