--- packages/plugin-react-native-span-access/ios/BugsnagJavascriptSpanControl+Private.h
@@ -1,7 +1,0 @@
-#import "BugsnagJavascriptSpanControl.h"
-
-@interface BugsnagJavascriptSpanControl ()
-
-- (instancetype)initWithSpanName:(NSString *)spanName;
-
-@end
