--- packages/plugin-react-native-span-access/ios/BugsnagJavascriptSpanControl.h
@@ -1,31 +1,0 @@
-#import <Foundation/Foundation.h>
-#import <BugsnagPerformance/BugsnagPerformanceSpanControl.h>
-#import <BugsnagPerformance/BugsnagPerformanceSpanContext.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-typedef void (^OnSpanUpdatedCallback)(BOOL result);
-
-typedef void (^RemoteSpanContextCallback)(BugsnagPerformanceSpanContext * _Nullable context);
-
-@interface BugsnagJavascriptSpanTransaction : NSObject
-
-- (void)end;
-
-- (void)endWithEndTime:(NSDate *)endTime NS_SWIFT_NAME(end(endTime:));
-
-- (void)setAttribute:(NSString *)attributeName withValue:(_Nullable id)value;
-
-- (void)commit:(OnSpanUpdatedCallback)callback;
-
-@end
-
-@interface BugsnagJavascriptSpanControl : NSObject<BugsnagPerformanceSpanControl>
-
-- (BugsnagJavascriptSpanTransaction *)createUpdateTransaction;
-
-- (void)retrieveSpanContext:(RemoteSpanContextCallback)callback;
-
-@end
-
-NS_ASSUME_NONNULL_END
