--- packages/plugin-react-native-span-access/ios/BugsnagJavascriptSpanControlProvider.h
@@ -1,10 +1,0 @@
-#import <Foundation/Foundation.h>
-#import <BugsnagPerformance/BugsnagPerformanceSpanControlProvider.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface BugsnagJavascriptSpanControlProvider: NSObject <BugsnagPerformanceSpanControlProvider>
-
-@end
-
-NS_ASSUME_NONNULL_END
