--- packages/plugin-react-native-span-access/ios/BugsnagJavascriptSpanQuery.h
@@ -1,12 +1,0 @@
-#import <BugsnagPerformance/BugsnagPerformanceSpanQuery.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface BugsnagJavascriptSpanQuery : BugsnagPerformanceSpanQuery
-
-+ (instancetype)queryWithName:(nonnull NSString *)name;
-
-@end
-
-NS_ASSUME_NONNULL_END
-
