--- packages/plugin-react-native-span-access/ios/BugsnagJavascriptSpansPlugin.h
@@ -1,10 +1,0 @@
-#import <Foundation/Foundation.h>
-#import <BugsnagPerformance/BugsnagPerformancePlugin.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface BugsnagJavascriptSpansPlugin: NSObject <BugsnagPerformancePlugin>
-
-@end
-
-NS_ASSUME_NONNULL_END
