+++ packages/plugin-react-native-span-access/ios/BugsnagNativeSpans.h
@@ -1,5 +1,4 @@
 #import <React/RCTBridgeModule.h>
-#import <React/RCTEventEmitter.h>
 
 #ifdef RCT_NEW_ARCH_ENABLED
 #import <BugsnagNativeSpansSpec/BugsnagNativeSpansSpec.h>
@@ -7,7 +6,7 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface BugsnagNativeSpans : RCTEventEmitter <RCTBridgeModule>
+@interface BugsnagNativeSpans : NSObject <RCTBridgeModule>
 
 @end
