+++ test/react-native/features/fixtures/scenario-launcher/android/src/newarch/java/com/bugsnag/reactnative/scenariolauncher/ScenarioLauncher.java
@@ -4,7 +4,6 @@ import androidx.annotation.NonNull;
 import com.bugsnag.reactnative.scenariolauncher.NativeScenarioLauncherSpec;
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.WritableMap;
 
@@ -60,16 +59,6 @@ public class ScenarioLauncher extends NativeScenarioLauncherSpec {
   @Override
   public void endNativeSpan(String traceParent, Promise promise) {
     impl.endNativeSpan(traceParent, promise);
-  }
-
-  @Override
-  public void updateJavascriptSpan(String spanName, ReadableArray attributes, Promise promise) {
-    impl.updateJavascriptSpan(spanName, attributes, promise);
-  }
-
-  @Override
-  public void sendNativeSpanWithJsParent(String spanName, Promise promise) {
-    impl.sendNativeSpanWithJsParent(spanName, promise);
   }
 }
