+++ test/react-native/features/fixtures/scenario-launcher/ios/ScenarioLauncher.mm
@@ -5,11 +5,7 @@
 #ifdef NATIVE_INTEGRATION
 #import <BugsnagPerformance/BugsnagPerformance.h>
 #import <BugsnagPerformance/BugsnagPerformanceConfiguration+Private.h>
-#import <BugsnagPerformance/BugsnagPerformanceSpanContext.h>
 #import "BugsnagNativeSpansPlugin.h"
-#import "BugsnagJavascriptSpansPlugin.h"
-#import "BugsnagJavascriptSpanQuery.h"
-#import "BugsnagjavascriptSpanControl.h"
 #endif
 
 @implementation ScenarioLauncher {
@@ -149,7 +145,6 @@ RCT_EXPORT_METHOD(startNativePerformance:(NSDictionary *)configuration resolve:(
     config.internal.clearPersistenceOnStart = YES;
 
     [config addPlugin:[BugsnagNativeSpansPlugin new]];
-    [config addPlugin:[BugsnagJavascriptSpansPlugin new]];
 
     [BugsnagPerformance startWithConfiguration:config];
     resolve(nil);
@@ -180,47 +175,6 @@ RCT_EXPORT_METHOD(endNativeSpan:(NSString *)traceParent resolve:(RCTPromiseResol
   }
 }
 
-RCT_EXPORT_METHOD(updateJavascriptSpan:(NSString *)spanName attributes:(NSArray<NSDictionary *> *)attributes resolve:(RCTPromiseResolveBlock)resolve reject:(RCTPromiseRejectBlock)reject) {
-  BugsnagJavascriptSpanQuery *spanQuery = [BugsnagJavascriptSpanQuery queryWithName:spanName];
-  BugsnagJavascriptSpanControl *spanControl = [BugsnagPerformance getSpanControlsWithQuery:spanQuery];
-  BugsnagJavascriptSpanTransaction *transaction = [spanControl createUpdateTransaction];
-  NSDate *endTime = [NSDate date];
-  
-  // Set some attributes
-  for (NSDictionary *attribute in attributes) {
-    NSString *name = attribute[@"name"];
-    id value = attribute[@"value"];
-    if (name) {
-      [transaction setAttribute:name withValue:value];
-    }
-  }
-
-  // End the span with a specific end time and commit the transaction
-  [transaction endWithEndTime:endTime];
-  [transaction commit:^(BOOL result) {
-    NSLog(@"Span update result: %@", result ? @"YES" : @"NO");
-  }];
-
-  resolve(nil);
-}
-
-RCT_EXPORT_METHOD(sendNativeSpanWithJsParent:(NSString *)spanName resolve:(RCTPromiseResolveBlock)resolve reject:(RCTPromiseRejectBlock)reject) {
-  BugsnagJavascriptSpanQuery *spanQuery = [BugsnagJavascriptSpanQuery queryWithName:spanName];
-  BugsnagJavascriptSpanControl *spanControl = [BugsnagPerformance getSpanControlsWithQuery:spanQuery];
-
-  [spanControl retrieveSpanContext:^(BugsnagPerformanceSpanContext * _Nullable context) {
-    if (context) {
-      BugsnagPerformanceSpanOptions *options = [BugsnagPerformanceSpanOptions new];
-      options.parentContext = context;
-      BugsnagPerformanceSpan *childSpan = [BugsnagPerformance startSpanWithName:@"Native Child Span" options:options];
-      [childSpan end];
-      resolve(nil);
-    } else {
-      reject(@"send_native_span_with_js_parent", @"Failed to get JS span context", nil);
-    }
-  }];
-}
-
 #else
 RCT_EXPORT_METHOD(startNativePerformance:(NSDictionary *)configuration resolve:(RCTPromiseResolveBlock)resolve reject:(RCTPromiseRejectBlock)reject) {
   NSLog(@"Native performance is not enabled in this build");
@@ -235,16 +189,6 @@ RCT_EXPORT_METHOD(startNativeSpan:(NSDictionary *)options resolve:(RCTPromiseRes
 RCT_EXPORT_METHOD(endNativeSpan:(NSString *)traceParent resolve:(RCTPromiseResolveBlock)resolve reject:(RCTPromiseRejectBlock)reject) {
   NSLog(@"Native performance is not enabled in this build");
   reject(@"end_native_span", @"Native performance is not enabled in this build", nil);
-}
-
-RCT_EXPORT_METHOD(updateJavascriptSpan:(NSString *)spanName attributes:(NSArray<NSDictionary *> *)attributes resolve:(RCTPromiseResolveBlock)resolve reject:(RCTPromiseRejectBlock)reject) {
-  NSLog(@"Native performance is not enabled in this build");
-  reject(@"update_javascript_span", @"Native performance is not enabled in this build", nil);
-}
-
-RCT_EXPORT_METHOD(sendNativeSpanWithJsParent:(NSString *)spanName resolve:(RCTPromiseResolveBlock)resolve reject:(RCTPromiseRejectBlock)reject) {
-  NSLog(@"Native performance is not enabled in this build");
-  reject(@"send_native_span_with_js_parent", @"Native performance is not enabled in this build", nil);
 }
 
 #endif
