+++ test/react-native/features/fixtures/scenario-launcher/lib/NativeScenarioLauncher.ts
@@ -9,8 +9,6 @@ export interface Spec extends TurboModule {
   startNativePerformance(configuration: UnsafeObject): Promise<void>;
   startNativeSpan(options: UnsafeObject): Promise<string>;
   endNativeSpan(traceParent: string): Promise<boolean>;
-  updateJavascriptSpan(spanName: string, attributes: Array<{ [key: string]: any }>): Promise<void>;
-  sendNativeSpanWithJsParent(spanName: string): Promise<void>;
 }
 
 export default TurboModuleRegistry.get<Spec>("ScenarioLauncher") as Spec | null;
