+++ test/react-native/features/fixtures/scenario-launcher/scenarios/native-integration/index.js
@@ -3,6 +3,4 @@ export * as NativeIntegrationJsParentScenario from './NativeIntegrationJsParentS
 export * as NativeIntegrationNativeParentScenario from './NativeIntegrationNativeParentScenario'
 export * as RemoteParentContextNativeScenario from './RemoteParentContextNativeScenario'
 export * as RemoteParentContextJSScenario from './RemoteParentContextJSScenario'
-export * as NativeSpansPluginScenario from './NativeSpansPluginScenario'
-export * as JavascriptSpansPluginScenario from './JavascriptSpansPluginScenario'
-export * as JavascriptSpansContextScenario from './JavascriptSpansContextScenario'
+export * as RemoteSpansScenario from './RemoteSpansScenario'
